Extract ListRenderer component
